Add files via upload
This template will create a traditional 3 tier architecture in AWS. 
As of now it creates 1 public subnet, 1 Internet Gateway, 1 public route. Also, creates 1 private subnet for shared resources such as EC2, and 1 private subnet intended to be used for database.
The nomenclature of the resources were created with "dev" prefix as it's intended for a development environment.